complete variant

+++ BackEnd/app/Http/Controllers/Admin/ColorController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Color;
+use App\Models\Variant;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -31,7 +32,7 @@ class ColorController extends Controller
         $color->name = $request->input('name');
         $color->description = $request->input('description');
         $color->save();
-        return redirect()->route('admin.color.index')->with(['msg' => 'Màu đã được thêm.']);
+        return redirect()->route('admin.color.index')->with(['success' => 'Màu đã được thêm.']);
     }
     public function edit(Request $request, string $id)
     {
@@ -51,17 +52,20 @@ class ColorController extends Controller
             // Lưu thay đổi vào cơ sở dữ liệu
             $color->save();
         } else {
-            return redirect()->route('admin.color.index')->with(['msg' => 'Có gì đó không đúng! Xin thử lại']);
+            return redirect()->route('admin.color.index')->with(['error' => 'Có gì đó không đúng! Xin thử lại']);
         }
-        return redirect()->route('admin.color.index')->with(['msg' => 'Sửa thành công']);
+        return redirect()->route('admin.color.index')->with(['success' => 'Sửa thành công']);
     }
     public function destroy($id)
     {
         if ($id) {
-
+            $exists = Variant::where('color_id', $id)->exists();
+            if ($exists) {
+                return redirect()->route('admin.color.index')->with(['error' => 'Màu này đang tồn tại trong sản phẩm, lên không thể xoá!']);
+            }
             DB::table('color')->where('id', $id)->delete();
 
-            return redirect()->route('admin.color.index')->with(['msg' => 'Xoá thành công']);
+            return redirect()->route('admin.color.index')->with(['success' => 'Xoá thành công']);
         }
     }
 }
